refactor: printResult 메서드 매개변수 타입 변경

+++ src/main/java/calculator/view/OutputView.java
@@ -1,8 +1,10 @@
 package calculator.view;
 
+import java.math.BigDecimal;
+
 public class OutputView {
 
-    public void printResult(int sum) {
+    public void printResult(BigDecimal sum) {
         System.out.println("결과 : " + sum);
     }
 }
